Taking Parameter in for signing requests
